feat: Automatically choose name to display
Some users might not have display names set, and all bots do not have a display name. This commit automatically chooses a non-empty name to use as the link's name.

`display_name` is always prioritized and is used if non-empty; otherwise (for users without a display name or bots), `real_name` is used.